"wip : refactor injective types and templates"
