+++ packages/plugin-injective/injective-sdk-client-ts/src/types/auction.ts
@@ -1,0 +1,19 @@
+//Auction Module Params
+export interface MsgBidRequestParams {
+    amount: string;
+}
+export interface GetAuctionRoundParams {
+    round: number;
+}
+
+export interface GetAuctionsParams {
+    startRound: number;
+    limit: number;
+}
+
+export interface MsgExternalTransferParams {
+    srcSubaccountId: string;
+    dstSubaccountId: string;
+    tokenSymbol: string;
+    amount: string;
+}
