+++ packages/plugin-injective/injective-sdk-client-ts/src/types/peggy.ts
@@ -1,0 +1,12 @@
+
+//Peggy
+export interface MsgSendToEthParams {
+    amount: {
+        denom: string;
+        amount: string;
+    };
+    bridgeFee?: {
+        denom: string;
+        amount: string;
+    };
+}
